about page updated with image

+++ Desktop/Development/milestone-11-getting-started-with-node-mongo/Assignment/travel-tour-react/src/components/Pages/DeleteOffer/DeleteOffer.js
@@ -12,17 +12,21 @@ const DeleteOffer = () => {
         }, []);
 
     const del = <FontAwesomeIcon icon={faTrashAlt} className="text-danger"/>;
-    const handleDelete = (id) => {
-        fetch(`https://infinite-stream-42915.herokuapp.com/offerings/${id}`, {
-          method: "DELETE",
-        })
-          .then((res) => res.json())
-          .then((data) => {
-            if (data.deletedCount > 0) {
-              const rest = orders.filter((order) => order._id !== id);
-              setOrders(rest);
-            }
-          });
+  const handleDelete = (id) => {
+    const proceed = window.confirm('Confirm delete?');
+    if (proceed) {
+      
+      fetch(`https://infinite-stream-42915.herokuapp.com/offerings/${id}`, {
+        method: "DELETE",
+      })
+        .then((res) => res.json())
+        .then((data) => {
+          if (data.deletedCount > 0) {
+            const rest = orders.filter((order) => order._id !== id);
+            setOrders(rest);
+          }
+        });
+    }
         
         
       };
